requirements.txt
anyio==3.6.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.89.1
flake8==6.0.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
protobuf==3.20.3
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.22.0
twilio==7.16.4
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
